Refactor out all of the card handling to a shared component
