+++ src/components/layouts/LayoutCorners/cardSwitcher.scss
@@ -1,5 +1,5 @@
 .card-icon-container {
-  z-index: 10000 !important;
+  z-index: 1000000 !important;
   width: 80px;
   position: absolute;
   top: 320px;
@@ -7,7 +7,8 @@
   z-index: 10;
   &:hover {
     .card-icon-item {
-      .card-button-background, .btn {
+      .card-button-background,
+      .btn {
         left: 5px;
       }
       .card-button-mask {
@@ -116,7 +117,7 @@
       transition: width 0.5s ease-in-out 0.2s;
       z-index: -1;
     }
-   /* .card-button {
+    /* .card-button {
       font-weight: 700;
       text-transform: capitalize;
       height: 40px;
