+++ src/components/layouts/LayoutGlass/index.js
@@ -1,34 +1,16 @@
 import React, { Component } from "react";
-import Views from "../../views";
 import Alerts from "../../generic/Alerts";
 import ActionsMixin from "../../generic/Actions";
 import TransitionGroup from "react-addons-transition-group";
-import CardHolder from "./cardHolder";
 import CardFrame from "./frame";
 import Widgets from "../LayoutOdyssey/widgets";
 import { withApollo } from "react-apollo";
-import gql from "graphql-tag";
 import CardSwitcher from "../LayoutCorners/CardSwitcher";
+import renderCards from "../cardRenderer";
 import "./layout.css";
 
 class LayoutGlass extends Component {
   state = {};
-  stopTraining = () => {
-    const client = this.props.clientObj.id;
-    const variables = {
-      client,
-      training: false
-    };
-    const mutation = gql`
-      mutation ClientSetTraining($client: ID!, $training: Boolean!) {
-        clientSetTraining(client: $client, training: $training)
-      }
-    `;
-    this.props.client.mutate({
-      mutation,
-      variables
-    });
-  };
   render() {
     let {
       simulator,
@@ -38,53 +20,12 @@ class LayoutGlass extends Component {
       clientObj,
       flight
     } = this.props;
-    const { login: stationLogin, name: stationName } = station;
+    const { name: stationName } = station;
     let alertClass = `alertColor${simulator.alertlevel || 5}`;
-    if (clientObj.loginState === "logout" && stationLogin === false) {
-      cardName = "Login";
-    }
-    if (clientObj.offlineState) {
-      cardName = "Offline";
-    }
     return (
       <ActionsMixin {...this.props}>
         <div className={`layout-glass ${alertClass}`}>
-          <TransitionGroup>
-            {station.cards
-              .concat({ name: "Login", component: "Login", icon: "Login" })
-              .concat({
-                name: "Offline",
-                component: "Offline",
-                icon: "Offline"
-              })
-              .map(card => {
-                if (card.name === cardName) {
-                  const component = Views[card.component];
-                  if (card.component.match(/.{8}-.{4}-.{4}-.{4}-.{12}/gi)) {
-                    // Software Panel
-                    return (
-                      <CardHolder
-                        component={Views.SoftwarePanels}
-                        panel={card.component}
-                        {...this.props}
-                        stopTraining={this.stopTraining}
-                        key={card.name}
-                      />
-                    );
-                  }
-                  return (
-                    <CardHolder
-                      component={component}
-                      {...this.props}
-                      stopTraining={this.stopTraining}
-                      key={card.name}
-                    />
-                  );
-                }
-                return null;
-              })
-              .filter(card => card)}
-          </TransitionGroup>
+          <TransitionGroup>{renderCards(this.props)}</TransitionGroup>
 
           <div className="frame-text">
             <h1 className="simulator-name">{simulator.name}</h1>
