+++ src/components/layouts/LayoutJr/index.js
@@ -1,42 +1,16 @@
 import React from "react";
-import Views from "../../views";
 import Alerts from "../../generic/Alerts";
 import ActionsMixin from "../../generic/Actions";
+import renderCards from "../cardRenderer";
 import "./style.css";
 
 export default props => {
-  const { simulator, station, clientObj } = props;
-  let cardName = props.cardName;
+  const { simulator, station } = props;
   let alertClass = `alertColor${simulator.alertlevel || 5}`;
-  if (clientObj.loginState === "logout" && station.login === false) {
-    cardName = "Login";
-  }
-  if (clientObj.offlineState) {
-    cardName = "Offline";
-  }
   return (
     <ActionsMixin {...props}>
       <div className={`layout-jr card-area ${alertClass}`}>
-        {station.cards
-          .concat({ name: "Login", component: "Login", icon: "Login" })
-          .concat({
-            name: "Offline",
-            component: "Offline",
-            icon: "Offline"
-          })
-          .map(card => {
-            if (card.component.match(/.{8}-.{4}-.{4}-.{4}-.{12}/gi)) {
-              return (
-                <Views.SoftwarePanels panel={card.component} {...this.props} />
-              );
-            }
-            const Component = Views[card.component];
-            if (card.name === cardName) {
-              return <Component {...props} />;
-            }
-            return null;
-          })
-          .filter(card => card)}
+        {renderCards(props)}
       </div>
       <div id="jr-frame" className={alertClass}>
         <div className="title-bar" />
