+++ src/containers/Card.js
@@ -72,7 +72,7 @@ export default class CardFrame extends Component {
             id: "test",
             name: "Test",
             alertLevel: "5",
-            layout: "LayoutCorners"
+            layout: "LayoutGlass"
           },
           station: {
             name: "Test",
